\name{beta_density}
\alias{beta_density}
\title{
  Function for Beta PDF
}
\description{
This function takes argument shape1 and shape2 and computes a beta distribution based upon those variables.  In the function we use "dbeta" which is a built in function that actually computes the function.
}
\usage{
beta_density(x, shape1, shape2)
}
\arguments{
  \item{x}{
x means to plot the values which dbeta generates on the x-axis.
}
\item{shape1}{Shape1 refers to the commonly known parameter alpha.
}
  \item{shape2}{
Shape2 refers to the commonly known parameter beta.
}
}
