\name{beta_func}
\alias{beta_func}
\title{
Beta Moments Function
}
\description{
Function that contains the moment functions for Beta distributed random variables.  Uses the gmm() function to estimate sample parameters based on sample data.
\usage{
beta_density(theta, x)
}
\arguments{
  \item{theta}{
Theta is a vector of 2 parameters, commonly known as alpha and beta, from which the population mean calculated.  Alpha and Beta are the initial guess parameters from the user.
}
\item{x}{x refers to the vector of sample data, in this case a sample of beta distributed RVs.
}
  
}
