\name{beta_sim}
\alias{beta_sim}
\title{
Beta Data Simulation
}
\description{
Function  that generates n beta distributed random variables.
\usage{
beta_density(n, shape1, shape2)
}
\arguments{
  \item{n}{
n is the desired sample size of the user.
}
\item{shape1}{shape1 refers to the common parameter alpha that is used in generating the data.
}
\item{shape2}{
SHape2 is the common parameter beta.  
}

