\name{gamma_density}
\alias{gamma_density}
\title{
Gamma Density Function
}
\description{
Function  that is used to plot the theoretical gamma density function with the estimateed parameters.
\usage{
gamma_density{x, shape, rate
}
\arguments{
  \item{x}{
x is the sample data required to plot the function against.
}
\item{shape}{shape is the parameter alpha used to compute density values.
}
\item{rate}{
rate is the parameter alpha used to compute density values.
}
}

