\name{gamma_func}
\alias{gamma_func}
\title{
Gamma Distribution Moments Function
}
\description{
Function that is used to calculate 1st and second moments of gamma distribution and ffed to the gmm() function.
\usage{
gamma_func(theta, x){
}
\arguments{
  \item{theta}{
theta is the vector of initial guess parameters, alpha and beta.
}
\item{x}{x refers to the sample data used.
}

