\name{gamma_sim}
\alias{gamma_sim}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
gamma simulation function
}
\description{
The function will generate gamma simulated data
}
\usage{
gamma_sim(n, shape, rate)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{n}{
n is the size of data
}
  \item{shape}{
shape is a gamma distribution parameter
}
  \item{rate}{
rate is a gamma distribution parameter
}
}
