\name{poisMixture_density}
\alias{poisMixture_density}
\title{
poisson mixture density function
}
\description{
This function is poisson mixture density function.
}
\usage{
poisMixture_density(x, lambda1, lambda2, r)
}
\arguments{
  \item{x}{
x is the sample data
}
  \item{lambda1}{
lambda1 is the parameter of the first part of poisson mixture
}
  \item{lambda2}{
lambda1 is the parameter of the second part of poisson mixture
}
  \item{r}{
r is the parameter of proportion of poisson mixture
}
}
