\name{poisson2}
\alias{poisson2}
\title{
poisson mixture density function
}
\description{

}
\usage{
poisson2(lambda1, lambda2, r)
}
\arguments{
  \item{lambda1}{
  lambda1 represents the poisson parameter of the first part of poisson mixture
}
  \item{lambda2}{
lambda2 represents the poisson parameter of the second part of poisson mixture
}
  \item{r}{
r represents the porportion of the first part of poisson mixture
}
}

