\name{poisson_mixture_func}
\alias{poisson_mixture_func}
\title{
   poisson mixture moment function
}
\description{
   This function is the moment function of poisson mixture
}
\usage{
poisson_mixture_func(theta, x)
}
\arguments{
  \item{theta}{
theta is the vector of guess value
}
  \item{x}{
x is the simulated data
}
}

