\name{poisson_mixture_sim}
\alias{poisson_mixture_sim}
\title{
poission mixture simulation function
}
\description{
This function generates poisson mixture simulation data
}
\usage{
poisson_mixture_sim(n, lambda1, lambda2, r)
}
\arguments{
  \item{n}{
n is the size of data
}
  \item{lambda1}{
lambda1 is the parameter of the first part poisson
}
  \item{lambda2}{
lambda1 is the parameter of the second part poisson
}
  \item{r}{
r is the proportion of the first part poisson
}
}
