\name{power_law_func}
\alias{power_law_func}
\title{
   power law moment function
}
\description{
  This is the power law moment function
}
\usage{
power_law_func(theta, x)
}
\arguments{
  \item{theta}{
    theta is the initial guess values of the parameters
}
  \item{x}{
    x is the sample data
}
}
