\name{power_law_sim}
\alias{power_law_sim}
\title{
    power law simualtion function
}
\description{
This function generates the power law simulated data
}
\usage{
power_law_sim(n, xmin, k)
}
\arguments{
  \item{n}{
n is the size of simulated data
}
  \item{xmin}{
xmin is the lower bound
}
  \item{k}{
k is the estimated parameter for power law
}
}

