\name{powerlaw_density}
\alias{powerlaw_density}
\title{
 Power law density
}
\description{
  This function define the power density
}
\usage{
powerlaw_density(x, theta, c)
}
\arguments{
  \item{x}{
     x is the sample data
}
  \item{theta}{
     theta is the estimated parameter for the power law density
}
\item{c} {
     c is the scale parameter
}
}
